refactor: sample_report.py and couple_report.py are merged into a models file (models.py) (#34)
