+++ modules/report/__init__.py
@@ -1,5 +1,5 @@
 # modules/report/__init__.py
 
 from .report_manager import ReportManager
-from .sample_report import SampleReport, ReportTable
+from .models import SampleReport, ReportTable
 from .couple_report import CoupleReport
