+++ modules/report/models.py
@@ -1,4 +1,4 @@
-# modules/report/sample_report.py
+# modules/report/models.py
 
 from typing import List, Dict, Any
 
@@ -29,6 +29,27 @@ class SampleReport:
         self.tables: List[ReportTable] = []
 
     def add_table(self, table: ReportTable | None):
+        if table is None:
+            return
+        self.tables.append(table)
+
+class CoupleReport:
+    """
+    RR-only couple-level reproductive risk interpretation.
+    """
+
+    def __init__(self,
+                 sample_a_id: str,
+                 sample_b_id: str,
+                 rr_mode: str
+                 ):
+        self.sample_a_id = sample_a_id
+        self.sample_b_id = sample_b_id
+        self.rr_mode = rr_mode
+        self.tables: List[ReportTable] = []
+
+    def add_table(self, table: ReportTable | None):
+
         if table is None:
             return
         self.tables.append(table)
